financial_time_series/tensorflow-model/requirements.txt
google-cloud-storage==1.10.0
pandas
numpy
